requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
langchain>=0.0.200
langchain-community>=0.0.20
faiss-cpu>=1.7.4
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
scikit-image>=0.20.0